more opcodes

+++ Opcode.cpp
@@ -8,6 +8,6 @@ Opcode::Opcode(uint16_t opcode)
 	, m_y{ static_cast<uint8_t>((opcode & 0x00F0) >> 4) }
 	, m_n { static_cast<uint8_t>(opcode & 0x000F)}
 	, m_nn{ static_cast<uint8_t>(opcode & 0x00FF) }
-	, m_nnn { static_cast<uint8_t>(opcode & 0x0FFF)}
+	, m_nnn { static_cast<uint16_t>(opcode & 0x0FFF)}
 {
 }
